requirements.txt
python-telegram-bot
mysql-connector-python
cryptography
requests
schedule
flask